beautiful

+++ include/jevmachopp/CpuTypeMeta.h
@@ -16,16 +16,14 @@ static_assert_cond(sizeof(CpuTypeMeta) == sizeof(CpuType) + sizeof(CpuSubType));
 
 template <> struct fmt::formatter<CpuTypeMeta> {
 
-    template <typename ParseContext> constexpr auto parse(ParseContext& ctx)
-    {
+    template <typename ParseContext> constexpr auto parse(ParseContext &ctx) {
         return ctx.begin();
     }
 
-    template <typename FormatContext> auto format(CpuTypeMeta const& metatype, FormatContext& ctx)
-    {
-        return fmt::format_to(ctx.out(),
-            "<CpuTypeMeta type: {:s} ({:#08x}) subtype: {:s} ({:#08x})>",
-            magic_enum::enum_name(metatype.type), to_underlying_int(metatype.type),
-            magic_enum::enum_name(metatype.subtype), to_underlying_int(metatype.subtype));
+    template <typename FormatContext> auto format(CpuTypeMeta const &metatype, FormatContext &ctx) {
+        return fmt::format_to(
+            ctx.out(), "<CpuTypeMeta type: {:s} ({:#08x}) subtype: {:s} ({:#08x})>",
+            magic_enum::enum_name(metatype.type), (uint32_t)to_underlying_int(metatype.type),
+            magic_enum::enum_name(metatype.subtype), (uint32_t)to_underlying_int(metatype.subtype));
     }
 };
